pyproject.toml
[project]
name = "valor"
version = "0.1.0"
description = "Code for the paper 'No Labels, No Problem: Training Visual Reasoners with Multimodal Veri"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "transformers==4.52.4",
    "ninja==1.11.1.4",
    "triton==3.2.0",
    "xformers==0.0.29.post2",
    "accelerate==1.10.0",
    "datasets==4.0.0",
    "deepspeed==0.18.2",
    "huggingface-hub==0.34.4",
    "hydra-core==1.3.2",
    "peft==0.17.0",
    "ray==2.48.0",
    "addict>=2.4.0",
    "colorlog>=6.10.1",
    "cython>=3.2.2",
    "fastapi==0.116.1",
    "google-genai==1.33.0",
    "huggingface-hub==0.34.4",
    "hydra-core==1.3.2",
    "iopath==0.1.10",
    "jsonlines>=4.0.0",
    "numpy==2.2.6",
    "openai==1.98.0",
    "opencv-python==4.12.0.88",
    "pandas==2.3.2",
    "pillow==11.3.0",
    "pip>=25.3",
    "pycocotools>=2.0.10",
    "pyyaml>3.10",
    "safetensors==0.6.2",
    "scipy>=1.15.3",
    "setuptools>=80.9.0",
    "submitit>=1.5.3",
    "supervision==0.6.0",
    "termcolor>=3.2.0",
    "timm==1.0.19",
    "tqdm==4.67.1",
    "yapf==0.40.1",
    "jupyterlab>=4.5.0",
    "ipykernel>=7.1.0",
]

[tool.uv]
override-dependencies = ["numpy==2.2.6","transformers==4.52.4"]

[tool.uv.pip]
no-build-isolation-package = ["groundingdino"]
